Fix go tool cover running from outside of gopath on files inside it.
The go cover file has always been a bit weird in my opinion when it
states the package path rather than the file path. In this case that
fails if we run go 1.11 from outside $GOPATH/src since it will assume
that modules are in use then, even though they're not.

We might want to require modules later, but I don't think we'll do that
before go 1.12 is released so for now disable the modules when we detect
this condition.