Skip GNU ld flags and objcopy usage on OS X. (#18)
The --unresolved-symbols flags appears to be unneeded. I haven't been
able to find a proper way to disable the automatic loading of the Go
runtime on OS X. There might be a way but the symbol manipulation tools
seems slightly more limited than for ELF, so I'm not sure. I'm disabling
it for now (meaning you can't fork if you mix C and Go code).