Add support for passing flags to go build, e.g. -tags.
At first I simply used the environment variable, but this does not work
very well in a build system. Instead use those as a default but allow
them to be set statically in a configvars file. Then users can choose if
they want to depend on the environment or use a specific value.