Allow to add configvars files on command line.
This can be used to temporarily alter to compilation and enable e.g.
special go build tags. Used in sebase to enable to Go code calling cgo.

This should be seen as a temporary measure though since you have to use
it every time you call seb. The proper way is to add them in CONFIG.