Only use custom pkgdir for custom build modes.
The default one should use the default go packages part of the system
installation instead.

This reverts #13 and brings back the problem of downloading modules in
parallell for Go 1.11. I won't fix that as it's solved in Go 1.12 and
also go mod download works as a workaround, as well as not using go
modules.