Rework plugin support so it works on go 1.13.
Go 1.13 doesn't like the temporary directories, and they were always
quite the hack anyway. Rework it so it simply loads the compiled dynamic
objects.

This is a backwards incompatible change, because the plugin must now be
a main package rather than anything else. But I don't think there's any
other easy way to support plugins on 1.13 so I will go for this. It's
more evidence that the next release should probably be called 2.0
regardless of its size.