Add a godeps_rule CONFIG argument to set the rule for godeps.
I was hoping to do this with a variable rather than a custom argument,
but I couldn't get it to work, it seems ninja doesn't allow variables
for the rule name.

While touch should be the default rule it wouldn't be backwards
compatible so keeping the godeps default for now.