Add support for go modules/plugins.
As we call the similar C binary a module I went with that name for the
Go version as well. There's some trickery needed, Go does currently not
support nocgo for modules, as noted in
https://github.com/golang/go/issues/30986
so that functionality has to be disabled. Also this triggers the CC
environment variable problem again so I had to change the test CC value
used.