Rewrite gobuild as a Go tool.
Invoked with seb -tool gobuild. I guess the diff stat is a testament to
how expressing bash is. The new code should be more readable however,
having fewer binaries also help. I have an idea that sebuild should be
self-contained in a single binary eventually and this is a step in that
direction.