Allow multiple targets in specialsrcs.
Quite a few commands generate multiple output and this should make those
easier to work with, as long as they want the list on the commandline.

Since my first implementation ended up with an unstable order in the
$out variable I added a test for that.

The temporary target renaming interfered a bit, but I think I got the
parts right. The rename check is still done for each target name added,
even when they're part of a multi-target.