Add a stamp file and a touch tool.
It's sometimes useful to set an mtime to the future to avoid rebuilding
something until that timestamp passes. Typically needed when a file has
a builtin expiration timestamp such as a certificate.

For this to be possible sebuild now always creates a stamp file in
$buildpath with the current timestamp. The file can then use this as a
source or dependency to always be rebuilt when it's mtime passes the
current time.

Setting the timestamp to the future turned out to be difficult with the
defaul touch tool, there's no portable way. Because of that I decided to
write an internal tool for it, which turned out to be pretty trivial.