Exclude files in goroot and imported modules from depfile. (#33)
While the goroot files change if you upgrade the Go version, there's
just so many of them that depending on them all is affecting seb
runtime. Thus I think it's a good trade-off to not put them in the
depfile.

I explored skipping the go list command and use build.Import directly,
but it turns out this runs go list regardless, so it was mostly
pointless and much slower.