Make the invars generation a builtin tool.
This requires a bit of change in how it's run. It's still a shell script
but I concatenate it with all the configvars files and buildvars and
pipe it to bash.

Thus including the parent invars is no longer required, instead you can
specify multiple invars in CONFIG and they'll all run.