By default, don't track Go dependencies.
Running the go list command to list dependencies are often much slower
than just re-running the compilation. This didn't use to be the case
before go modules, but since their introduction and the addition of the
Go build cache it very much is. It doesn't make sense to track Go
dependencies any longer since it's mostly handled internally by the go
tool.

Still keeping it as an option as disabling it makes ninja re-run stuff
on every run and it can't report "nothing to do" anylonger.